+++ src/main/java/structured/EuroToPeseta.java
@@ -7,8 +7,8 @@ public class EuroToPeseta {
         Scanner sc = new Scanner(System.in);
 
         System.out.println("Escribe la cantidad de euros a convertir");
-        double euros = sc.nextDouble();
-        double pesetas = euros*166.387;
+        float euros = sc.nextFloat();
+        float pesetas = euros*166.387f;
         System.out.println("La cantidad de pesetas para " +euros +" euros es " +pesetas);
     }
 }
