+++ src/main/java/structured/KilometersToMiles.java
@@ -6,8 +6,8 @@ public class KilometersToMiles {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Entra el número de kilometros");
-        double kilometers = sc.nextDouble();
-        double miles = kilometers/1.852;
+        float kilometers = sc.nextFloat();
+        float miles = kilometers/1.852f;
         System.out.println("El numero de kilometros para " +kilometers +"es: " +miles);
     }
 }
