+++ src/main/java/structured/MilesToKilometers.java
@@ -6,8 +6,8 @@ public class MilesToKilometers {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         System.out.println("Entra el número de millas");
-        double miles = sc.nextDouble();
-        double kiloneters = miles*1.852;
+        float miles = sc.nextFloat();
+        float kiloneters = miles*1.852f;
         System.out.println("El numero de kilometros para " +miles +"es: " +kiloneters);
     }
     }
